2 - RTB/PoC/backend/requirements.txt
Flask
flask_cors
pinecone-client
langchain
pypdf
PyPDF2
python-dotenv
openai
tiktoken
huggingface-hub
InstructorEmbedding
sentence-transformers
boto3
